updated constructor kit view with a list view and sub view per section to keep things cleaner
